@@ -1,23 +1,17 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <OutputType>Exe</OutputType>
-        <TargetFramework>net6.0</TargetFramework>
-        <RollForward>Major</RollForward>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <!-- CDK Construct Library dependencies -->
-        <PackageReference Include="Amazon.CDK.Lib" Version="2.167.2" />
-        <PackageReference Include="Cdklabs.CdkNag" Version="2.34.2" />
-        <PackageReference Include="Constructs" Version="10.4.2" />
-
-        <PackageReference Include="Amazon.Jsii.Analyzers" Version="*" PrivateAssets="all" />
-        
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\Bookstore.Common\Bookstore.Common.csproj" />
-    </ItemGroup>
-
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net8.0</TargetFramework>
+    <RollForward>Major</RollForward>
+  </PropertyGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Bookstore.Common\Bookstore.Common.csproj" />
+  </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="Amazon.CDK.Lib" Version="2.167.2" />
+    <PackageReference Include="Constructs" Version="10.4.2" />
+    <PackageReference Include="Cdklabs.CdkNag" Version="2.34.2" />
+    <PackageReference Include="Amazon.Jsii.Analyzers" Version="*" />
+  </ItemGroup>
 </Project>
