@@ -1,9 +1,9 @@
-﻿using Bookstore.Domain.Books;
+using Bookstore.Domain.Books;
 using Bookstore.Domain.Offers;
 using Bookstore.Domain.Orders;
 using Bookstore.Web.Areas.Admin.Models.Dashboard;
 using System.Threading.Tasks;
-using System.Web.Mvc;
+using Microsoft.AspNetCore.Mvc;
 
 namespace Bookstore.Web.Areas.Admin.Controllers
 {
@@ -12,7 +12,6 @@ namespace Bookstore.Web.Areas.Admin.Controllers
         private readonly IOrderService orderService;
         private readonly IOfferService offerService;
         private readonly IBookService bookService;
-
         public DashboardController(IOrderService orderService, IOfferService offerService, IBookService bookService)
         {
             this.orderService = orderService;
@@ -25,23 +24,19 @@ namespace Bookstore.Web.Areas.Admin.Controllers
             var orderStats = await orderService.GetStatisticsAsync();
             var offerStats = await offerService.GetStatisticsAsync();
             var inventoryStats = await bookService.GetStatisticsAsync();
-
             var model = new DashboardIndexViewModel
             {
                 PastDueOrders = orderStats.PastDueOrders,
                 PendingOrders = orderStats.PendingOrders,
                 OrdersThisMonth = orderStats.OrdersThisMonth,
                 OrdersTotal = orderStats.OrdersTotal,
-
                 PendingOffers = offerStats.PendingOffers,
                 OffersThisMonth = offerStats.OffersThisMonth,
                 OffersTotal = offerStats.OffersTotal,
-
                 LowStock = inventoryStats.LowStock,
                 OutOfStock = inventoryStats.OutOfStock,
                 StockTotal = inventoryStats.StockTotal
             };
-
             return View(model);
         }
     }
