@@ -1,5 +1,5 @@
-﻿
-using System.Web.Mvc;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Authorization;
 
 namespace Bookstore.Web.Areas.Admin.Controllers
 {
@@ -21,7 +21,6 @@ namespace Bookstore.Web.Areas.Admin.Controllers
             //var exception = HttpContext.Features.Get<IExceptionHandlerPathFeature>();
             //ViewData["Path"] = exception?.Path;
             //var error = Problem();
-
             //ViewData["StatusCode"] = error.StatusCode;
             return View("~/Views/Error/Index.cshtml");
         }
