@@ -1,12 +1,15 @@
-﻿using Bookstore.Domain.ReferenceData;
+using Bookstore.Domain.ReferenceData;
 using System.Collections.Generic;
-using System.Web.Mvc;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Rendering;
 
 namespace Bookstore.Web.Areas.Admin.Models.ReferenceData
 {
     public class ReferenceDataItemCreateUpdateViewModel
     {
-        public ReferenceDataItemCreateUpdateViewModel() { }
+        public ReferenceDataItemCreateUpdateViewModel()
+        {
+        }
 
         public ReferenceDataItemCreateUpdateViewModel(ReferenceDataItem referenceDataItem)
         {
@@ -16,11 +19,8 @@ namespace Bookstore.Web.Areas.Admin.Models.ReferenceData
         }
 
         public int Id { get; set; }
-
         public ReferenceDataType SelectedReferenceDataType { get; set; }
-
         public string Text { get; set; }
-
         public IEnumerable<SelectListItem> DataTypes { get; set; }
     }
 }
