@@ -1,6 +1,6 @@
-﻿using System;
-using System.Web;
-using System.Web.Mvc;
+using System;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Http;
 using BobsBookstoreClassic.Data;
 
 namespace Bookstore.Web.Controllers
@@ -9,8 +9,8 @@ namespace Bookstore.Web.Controllers
     {
         public ActionResult Login(string redirectUri = null)
         {
-            if(string.IsNullOrWhiteSpace(redirectUri)) return RedirectToAction("Index", "Home");
-
+            if (string.IsNullOrWhiteSpace(redirectUri))
+                return RedirectToAction("Index", "Home");
             return Redirect(redirectUri);
         }
 
@@ -21,9 +21,9 @@ namespace Bookstore.Web.Controllers
 
         private ActionResult LocalSignOut()
         {
-            if (HttpContext.Request.Cookies["LocalAuthentication"] != null)
+            if (Request.Cookies.ContainsKey("LocalAuthentication"))
             {
-                HttpContext.Response.Cookies.Add(new HttpCookie("LocalAuthentication") { Expires = DateTime.Now.AddDays(-1) });
+                Response.Cookies.Delete("LocalAuthentication");
             }
 
             return RedirectToAction("Index", "Home");
@@ -31,15 +31,14 @@ namespace Bookstore.Web.Controllers
 
         private ActionResult CognitoSignOut()
         {
-            if (Request.Cookies[".AspNet.Cookies"] != null)
+            if (Request.Cookies.ContainsKey(".AspNet.Cookies"))
             {
-                Response.Cookies.Add(new HttpCookie(".AspNet.Cookies") { Expires = DateTime.Now.AddDays(-1) });
+                Response.Cookies.Delete(".AspNet.Cookies");
             }
 
             var domain = BookstoreConfiguration.GetSetting("Authentication/Cognito/CognitoDomain");
             var clientId = BookstoreConfiguration.GetSetting("Authentication/Cognito/LocalClientId");
-            var logoutUri = $"{Request.Url.Scheme}://{Request.Url.Host}:{Request.Url.Port}/";
-
+            var logoutUri = $"{Request.Scheme}://{Request.Host}/";
             return Redirect($"{domain}/logout?client_id={clientId}&logout_uri={logoutUri}");
         }
     }
