@@ -1,9 +1,10 @@
-﻿using System.Threading.Tasks;
+using System.Threading.Tasks;
 using Bookstore.Web.Helpers;
 using Bookstore.Domain.Books;
 using Bookstore.Domain.Carts;
 using Bookstore.Web.ViewModel.Search;
-using System.Web.Mvc;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Authorization;
 
 namespace Bookstore.Web.Controllers
 {
@@ -12,7 +13,6 @@ namespace Bookstore.Web.Controllers
     {
         private readonly IBookService inventoryService;
         private readonly IShoppingCartService shoppingCartService;
-
         public SearchController(IBookService inventoryService, IShoppingCartService shoppingCartService)
         {
             this.inventoryService = inventoryService;
@@ -22,36 +22,28 @@ namespace Bookstore.Web.Controllers
         public async Task<ActionResult> Index(string searchString, string sortBy = "Name", int pageIndex = 1, int pageSize = 10)
         {
             var books = await inventoryService.GetBooksAsync(searchString, sortBy, pageIndex, pageSize);
-
             return View(new SearchIndexViewModel(books));
         }
 
         public async Task<ActionResult> Details(int id)
         {
             var book = await inventoryService.GetBookAsync(id);
-
             return View(new SearchDetailsViewModel(book));
         }
 
         public async Task<ActionResult> AddItemToShoppingCart(int bookId)
         {
             var dto = new AddToShoppingCartDto(HttpContext.GetShoppingCartCorrelationId(), bookId, 1);
-
             await shoppingCartService.AddToShoppingCartAsync(dto);
-
             this.SetNotification("Item added to shopping cart");
-
             return RedirectToAction("Index", "Search");
         }
 
         public async Task<ActionResult> AddItemToWishlist(int bookId)
         {
             var dto = new AddToWishlistDto(HttpContext.GetShoppingCartCorrelationId(), bookId);
-
             await shoppingCartService.AddToWishlistAsync(dto);
-
             this.SetNotification("Item added to wishlist");
-
             return RedirectToAction("Index", "Search");
         }
     }
