@@ -1,9 +1,10 @@
-﻿using System.Threading.Tasks;
+using System.Threading.Tasks;
 using Bookstore.Web.Helpers;
 using Bookstore.Domain.Customers;
 using Bookstore.Domain.Carts;
 using Bookstore.Web.ViewModel.ShoppingCart;
-using System.Web.Mvc;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Authorization;
 
 namespace Bookstore.Web.Controllers
 {
@@ -12,7 +13,6 @@ namespace Bookstore.Web.Controllers
     {
         private readonly ICustomerService customerService;
         private readonly IShoppingCartService shoppingCartService;
-
         public ShoppingCartController(ICustomerService customerService, IShoppingCartService shoppingCartService)
         {
             this.customerService = customerService;
@@ -22,7 +22,6 @@ namespace Bookstore.Web.Controllers
         public async Task<ActionResult> Index()
         {
             var shoppingCart = await shoppingCartService.GetShoppingCartAsync(HttpContext.GetShoppingCartCorrelationId());
-
             return View(new ShoppingCartIndexViewModel(shoppingCart));
         }
 
@@ -30,11 +29,8 @@ namespace Bookstore.Web.Controllers
         public async Task<ActionResult> Delete(int shoppingCartItemId)
         {
             var dto = new DeleteShoppingCartItemDto(HttpContext.GetShoppingCartCorrelationId(), shoppingCartItemId);
-
             await shoppingCartService.DeleteShoppingCartItemAsync(dto);
-
             this.SetNotification("Item removed from shopping cart.");
-
             return RedirectToAction(nameof(Index));
         }
 
