@@ -1,35 +1,27 @@
-﻿using Microsoft.Owin;
 using System;
 using System.Drawing;
 using System.Net;
-using System.Web;
+using Microsoft.AspNetCore.Http;
 
 namespace Bookstore.Web.Helpers
 {
     public static class HttpContextExtensions
     {
-        public static string GetShoppingCartCorrelationId(this HttpContextBase context)
+        public static string GetShoppingCartCorrelationId(this HttpContext context)
         {
             var CookieKey = "ShoppingCartId";
-
             var cookieOptions = new CookieOptions
             {
                 Expires = DateTime.Now.AddYears(1),
                 Path = "/"
             };
-
-            HttpCookie cookie = context.Request.Cookies[CookieKey];
-            string shoppingCartClientId = cookie != null ? cookie.Value : null;
-
-            //var shoppingCartClientId = context.Request.Cookies[CookieKey].Value;
-
+            string shoppingCartClientId = context.Request.Cookies[CookieKey];
             if (string.IsNullOrWhiteSpace(shoppingCartClientId))
             {
                 shoppingCartClientId = context.User.Identity.IsAuthenticated ? context.User.GetSub() : Guid.NewGuid().ToString();
             }
 
-            context.Response.Cookies.Add(new HttpCookie(CookieKey, shoppingCartClientId));
-
+            context.Response.Cookies.Append(CookieKey, shoppingCartClientId, cookieOptions);
             return shoppingCartClientId;
         }
     }
