@@ -1,5 +1,5 @@
-﻿using System.ComponentModel.DataAnnotations;
-using System.Web;
+using System.ComponentModel.DataAnnotations;
+using Microsoft.AspNetCore.Http;
 
 namespace Bookstore.Web.Helpers
 {
@@ -16,9 +16,9 @@ namespace Bookstore.Web.Helpers
         {
             if (value == null) return true;
 
-            if (!(value is HttpPostedFileBase file)) return base.IsValid(value);
+            if (!(value is IFormFile file)) return base.IsValid(value);
 
-            return file.ContentLength <= maxFileSize;
+            return file.Length <= maxFileSize;
         }
 
         public override string FormatErrorMessage(string name)
