@@ -1,9 +1,9 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Web;
-using System.Web.Mvc;
-using System.Web.SessionState;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Rendering;
+using Microsoft.AspNetCore.Mvc.ViewFeatures;
 
 namespace Bookstore.Web.Helpers
 {
@@ -19,6 +19,7 @@ namespace Bookstore.Web.Helpers
                     Text = emptyItem
                 };
             }
+
             foreach (var val in Enum.GetValues(typeof(T)))
             {
                 yield return new SelectListItem()
@@ -26,7 +27,6 @@ namespace Bookstore.Web.Helpers
                     Text = Enum.GetName(typeof(T), val)
                 };
             }
-
         }
     }
 }
