@@ -1,13 +1,16 @@
-﻿using Bookstore.Domain.ReferenceData;
+using Bookstore.Domain.ReferenceData;
 using System.Collections.Generic;
 using System.Linq;
-using System.Web.Mvc;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Rendering;
 
 namespace Bookstore.Web.ViewModel.Resale
 {
     public class ResaleCreateViewModel
     {
-        public ResaleCreateViewModel() { }
+        public ResaleCreateViewModel()
+        {
+        }
 
         public ResaleCreateViewModel(IEnumerable<ReferenceDataItem> referenceDataItems)
         {
@@ -18,27 +21,16 @@ namespace Bookstore.Web.ViewModel.Resale
         }
 
         public IEnumerable<SelectListItem> BookTypes { get; internal set; }
-
         public IEnumerable<SelectListItem> Publishers { get; internal set; }
-
         public IEnumerable<SelectListItem> Genres { get; internal set; }
-
         public IEnumerable<SelectListItem> Conditions { get; internal set; }
-
         public int SelectedBookTypeId { get; set; }
-
         public int SelectedPublisherId { get; set; }
-
         public int SelectedGenreId { get; set; }
-
         public int SelectedConditionId { get; set; }
-
         public decimal BookPrice { get; set; }
-
         public string BookName { get; set; }
-
         public string Author { get; set; }
-
         public string ISBN { get; set; }
     }
 }
