@@ -1,8 +1,8 @@
-﻿using System.Threading.Tasks;
-using System.Web.Mvc;
+using System.Threading.Tasks;
 using Bookstore.Domain.Offers;
 using Bookstore.Domain.ReferenceData;
 using Bookstore.Web.Areas.Admin.Models.Offers;
+using Microsoft.AspNetCore.Mvc;
 
 namespace Bookstore.Web.Areas.Admin.Controllers
 {
@@ -10,7 +10,6 @@ namespace Bookstore.Web.Areas.Admin.Controllers
     {
         private readonly IOfferService offerService;
         private readonly IReferenceDataService referenceDataService;
-
         public OffersController(IOfferService offerService, IReferenceDataService referenceDataService)
         {
             this.offerService = offerService;
@@ -21,7 +20,6 @@ namespace Bookstore.Web.Areas.Admin.Controllers
         {
             var offers = await offerService.GetOffersAsync(filters, pageIndex, pageSize);
             var referenceData = await referenceDataService.GetAllReferenceDataAsync();
-
             return View(new OfferIndexViewModel(offers, referenceData));
         }
 
@@ -52,11 +50,8 @@ namespace Bookstore.Web.Areas.Admin.Controllers
         private async Task<ActionResult> UpdateOfferStatus(int id, OfferStatus status, string message)
         {
             var dto = new UpdateOfferStatusDto(id, status);
-
             await offerService.UpdateOfferStatusAsync(dto);
-
             TempData["Message"] = message;
-
             return RedirectToAction("Index");
         }
     }
