@@ -1,10 +1,11 @@
-﻿using Bookstore.Domain.Addresses;
+using Bookstore.Domain.Addresses;
 using Bookstore.Domain.Carts;
 using Bookstore.Domain.Orders;
 using Bookstore.Web.Helpers;
 using Bookstore.Web.ViewModel.Checkout;
-using System.Web.Mvc;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.ModelBinding;
 
 namespace Bookstore.Web.Controllers
 {
@@ -13,10 +14,7 @@ namespace Bookstore.Web.Controllers
         private readonly IAddressService addressService;
         private readonly IShoppingCartService shoppingCartService;
         private readonly IOrderService orderService;
-
-        public CheckoutController(IShoppingCartService shoppingCartService,
-                                  IOrderService orderService,
-                                  IAddressService addressService)
+        public CheckoutController(IShoppingCartService shoppingCartService, IOrderService orderService, IAddressService addressService)
         {
             this.shoppingCartService = shoppingCartService;
             this.orderService = orderService;
@@ -27,26 +25,22 @@ namespace Bookstore.Web.Controllers
         {
             var shoppingCart = await shoppingCartService.GetShoppingCartAsync(HttpContext.GetShoppingCartCorrelationId());
             var addresses = await addressService.GetAddressesAsync(User.GetSub());
-
             return View(new CheckoutIndexViewModel(shoppingCart, addresses));
         }
 
         [HttpPost]
         public async Task<ActionResult> Index(CheckoutIndexViewModel model)
         {
-            if(!ModelState.IsValid) return  View(model);
-
+            if (!ModelState.IsValid)
+                return View(model);
             var dto = new CreateOrderDto(User.GetSub(), HttpContext.GetShoppingCartCorrelationId(), model.SelectedAddressId);
-
             var orderId = await orderService.CreateOrderAsync(dto);
-
             return RedirectToAction("Finished", new { orderId });
         }
 
         public async Task<ActionResult> Finished(int orderId)
         {
             var order = await orderService.GetOrderAsync(orderId);
-
             return View(new CheckoutFinishedViewModel(order));
         }
     }
