@@ -1,10 +1,10 @@
-﻿using Microsoft.Owin;
+using Microsoft.AspNetCore.Http;
 
 namespace Bookstore.Web.Helpers
 {
-    public static class OwinRequestExtensions
+    public static class HttpRequestExtensions
     {
-        public static string GetReturnUrl(this IOwinRequest request)
+        public static string GetReturnUrl(this HttpRequest request)
         {
             return $"{request.Scheme}://{request.Host}/signin-oidc";
         }
