@@ -1,74 +1,60 @@
-﻿using System;
-using Microsoft.Owin;
+using System;
 using System.Security.Claims;
 using System.Threading.Tasks;
-using System.Web;
 using Bookstore.Domain.Customers;
+using Microsoft.AspNetCore.Http;
 
 namespace Bookstore.Web.Helpers
 {
-    public class LocalAuthenticationMiddleware : OwinMiddleware
+    public class LocalAuthenticationMiddleware
     {
+        private readonly RequestDelegate _next;
         private const string UserId = "FB6135C7-1464-4A72-B74E-4B63D343DD09";
-
         private readonly ICustomerService _customerService;
 
-        public LocalAuthenticationMiddleware(OwinMiddleware next, ICustomerService customerService) : base(next)
+        public LocalAuthenticationMiddleware(RequestDelegate next, ICustomerService customerService)
         {
             _customerService = customerService;
+            _next = next;
         }
 
-        public override async Task Invoke(IOwinContext context)
+        public async Task InvokeAsync(HttpContext context)
         {
-            if (context.Request.Path.Value.StartsWith("/Authentication/Login"))
+            if (context.Request.Path.StartsWithSegments("/Authentication/Login"))
             {
                 CreateClaimsPrincipal(context);
-
-                await SaveCustomerDetailsAsync();
-
-                var userCookie = new HttpCookie("LocalAuthentication") { Expires = DateTime.Now.AddDays(1) };
-
-                HttpContext.Current.Response.Cookies.Add(userCookie);
-
+                await SaveCustomerDetailsAsync(context);
+                context.Response.Cookies.Append("LocalAuthentication", "true", new CookieOptions
+                {
+                    Expires = DateTime.Now.AddDays(1)
+                });
                 context.Response.Redirect("/");
+                return;
             }
-            else if (HttpContext.Current.Request.Cookies["LocalAuthentication"] != null)
+            else if (context.Request.Cookies.ContainsKey("LocalAuthentication"))
             {
                 CreateClaimsPrincipal(context);
-
-                await SaveCustomerDetailsAsync();
-
-                await Next.Invoke(context);
+                await SaveCustomerDetailsAsync(context);
             }
-            else
-            {
-                await Next.Invoke(context);
-            }
+
+            await _next(context);
         }
 
-        private void CreateClaimsPrincipal(IOwinContext context)
+        private void CreateClaimsPrincipal(HttpContext context)
         {
             var identity = new ClaimsIdentity("Application");
-
             identity.AddClaim(new Claim(ClaimTypes.Name, "bookstoreuser"));
             identity.AddClaim(new Claim("nameidentifier", UserId));
             identity.AddClaim(new Claim("given_name", "Bookstore"));
             identity.AddClaim(new Claim("family_name", "User"));
             identity.AddClaim(new Claim(ClaimTypes.Role, "Administrators"));
-
-            context.Request.User = new ClaimsPrincipal(identity);
+            context.User = new ClaimsPrincipal(identity);
         }
 
-        private async Task SaveCustomerDetailsAsync()
+        private async Task SaveCustomerDetailsAsync(HttpContext context)
         {
-            var identity = (ClaimsIdentity)HttpContext.Current.User.Identity;
-
-            var dto = new CreateOrUpdateCustomerDto(
-                identity.FindFirst("nameidentifier").Value,
-                identity.Name,
-                identity.FindFirst("given_name").Value,
-                identity.FindFirst("family_name").Value);
-
+            var identity = (ClaimsIdentity)context.User.Identity;
+            var dto = new CreateOrUpdateCustomerDto(identity.FindFirst("nameidentifier").Value, identity.Name, identity.FindFirst("given_name").Value, identity.FindFirst("family_name").Value);
             await _customerService.CreateOrUpdateCustomerAsync(dto);
         }
     }
